Removing un-necessary InternalVisibleTo I had added and using linked file for LanguageVersionExtensions
